try to fix ci
